+++ droit/plugins/output/math/main.py
@@ -1,5 +1,0 @@
-def calc(data, rpack):
-	equation = data[0].lower().replace("plus", "+").replace("minus", "-")
-	equation = equation.replace("^", "**").replace("x", "*").replace(",", ".")
-	return eval(equation), rpack
-