+++ sample/sample-bot-multi.py
@@ -1,16 +1,18 @@
-# sample-bot.py - a sample bot using python-droit
-
+# sample-bot-multi.py - a multi-session sample bot using python-droit
 import droit
 
-db = droit.Database()
+db = droit.Database(multiSession=True)
 db.loadPlugins()
-db.parseScript("german-sample.dda") # or using DroitXML: db.parseDroitXML("german-sample.xml")
+db.parseScript("sample/german-sample.dda")
+
+db.sessions.path = "sample/sessions.json"
+db.sessions.loadSessions()
 
 running = True
 
 while(running):
     try:
-        rawInput = input("Droit> ")
+        rawInput = input("(" + db.sessions.droitname + ") ")
         userinput = droit.models.DroitUserinput(rawInput) # Create DroitUserinput object
 
         hits = db.useRules(userinput) # Run userinput on database
@@ -22,7 +24,15 @@ while(running):
             output = db.formatOut(hit.rule.output, variables) #  Generate the output from using an output-rule
 
             print(output)
+
+            if(db.sessions.getActive()):
+                db.history.newEntry(userinput, hit.rule, output, userId=db.sessions.getActive().id)
+            else:
+                db.history.newEntry(userinput, hit.rule, output)
         else:
             print("Can't find an answer on this question.")
+            db.history.newEntry(userinput, None, None)
+    
     except KeyboardInterrupt:
+        db.sessions.saveSessions()
         running = False