+++ CHANGELOG.md
@@ -3,7 +3,47 @@ All notable changes to this project will be documented in this file.
 
 | Latest release | Version of master branch |  
 |----------------|--------------------------|
-| 1.0.4          | 1.0.4                    |
+| 1.0.3          | 1.1.0:6 beta             |
+
+
+## 1.1.0:4-6
+### Added
+- `SESSION` plugins added
+- `models.DroitSession` which contains a username and can store user data
+- `models.DroitMultiSession` which stores a list of `DroitSession` objects
+
+### Changed
+- `preloadScript` was added to input plugins
+- all parsing functions have the new optional parameter `append=False` which allows to append the parsed rules to the current rules
+
+### Removed
+- `parseDroitXML` and `writeDroitXML` were removed. However they will still be available in the legacy submodule (`legacy.parseDroitXML`)
+- `models.DroitSettings` which was also located within `Database.settings`
+
+## 1.1.0:1-3
+### Added
+- new DDS parsing algorithm used by the new functions `parseScript` and `parseScriptString`
+- `writeScriptString` makes it possible dump a Droit Database to a DDS string
+
+### Changed
+- `writeLegacy` method name was changed to `writeScript`
+- `DroitRuleInOut` was split up to `DroitRuleInput` and `DroitRuleOutput`
+- `legacy.checkValidLine()` improved
+- tests changed to use `parseScript`
+
+## 1.1.0 alpha
+Droit v1.1 and v1.0 are not compatible. The Database formats haven't changed but the library uses different commands and objects so you will have to rewrite you bot but you will still be able to use your old databases.
+### 1. Database object
+A new object called `Database` will be now the core of python-droit. It combines the functionallity of `__init__.py`, `tools.py`, `loader.py` and `dumper.py` from Droit v1.0. You can use it to parse and dump a Database, you can load plugins into it and run `useRules()` and `formatOut()` from it. It replaces `rpack` as it also stores `cache` and `history` objects. The `settings` object was removed and replaced by so called sessions.
+
+### 2. Plugins
+Restructuring droit of caurse has an effect on the plugins. There are minimal adjustments that have to be done to make v1.0 plugins work with v1.1. The biggest change is that there are now two types of output plugins. On the one hand there are still those that are called using `EVAL`. The folder they are stored in is from now on called `eval.pluginName`. The new type of output plugins is similar to input plugins. Those plugins use a `block()` function. The `TEXT` and `VAR` blocks are no longer part of the core source code but plugins.
+
+### 3. DroitIO
+DroitIO will be removed as it is the task of the person who writes the bot to implement a solution for input and output. Probably there will be a ready-to-use bot based on python-droit soon that can be configured easily similar to what DroitIO offered.
+
+### 4. Sessions
+Droit v1.1 is the first version to support multiple users. Therefore so called sessions can be created. The `Database` object stores a `DroitMultiSession` object in it's parameter `sessions`. Using this object you can load and save sessions and switch between users. Plugins can access sessions and store data separately for each user.
 
 ## 1.0.4 release
 ### Changed